issue-634: refactor(expenses): extract expense_tracker sub-widgets (#634)
Reduce expense_tracker_screen.dart 1466 -> 918 lines by extracting the
self-contained, presentational pieces into lib/widgets/expense/ (no
behavioural change):

- search_result_tile.dart      (SearchResultTile, verbatim)
- category_section.dart        (CategorySection, verbatim)
- expense_alerts_card.dart     (ExpenseAlertsCard)
- expense_recent_card.dart     (ExpenseRecentCard)
- expense_search_view.dart     (ExpenseSearchView — search bar + filter bar +
                                results; state stays in the screen, passed in)
- utils/expense_category_format.dart (localizedExpenseCategory helper)

The remaining >500 is the _buildNormalView body, whose extraction needs the
state-management layer from #632 (per #634's own migration note). Verified:
flutter analyze clean on changed files; flutter test test/screens/ green
(incl. expense_tracker_search_test.dart, which covers the extracted filter).

Part of #634 (expense tracker; meal_planner/dashboard/coach remain).
